# Added explicit register pressure estimate for SIMD and tuned [Dynamic]LinearQuantization operations